+++ R/MoEClust.R
@@ -4,9 +4,11 @@
 #' @section Usage:
 #' The most important function in the \pkg{MoEClust} package is: \code{\link{MoE_clust}}, for fitting the model via EM/CEM with gating and/or expert network covariates, supplied via formula interfaces.
 #'
-#' Other functions also exist, e.g. \code{\link{MoE_control}}, \code{\link{MoE_crit}}, \code{\link{MoE_dens}}, \code{\link{MoE_estep}}, \code{\link{MoE_compare}}, and \code{\link{aitken}}, which are all used within \code{\link{MoE_clust}} but are nonetheless made available for standalone use.
-#'
 #' \code{\link{MoE_compare}} is provided for conducting model selection between different results from \code{\link{MoE_clust}} using different covariate combinations &/or initialisation strategies, etc.
+#' 
+#' \code{\link{MoE_stepwise}} is provided for conducting a greedy forward stepwise search to identify the optimal model in terms of the number of components, GPCM covariance type, and the subsets of gating/expert network covariates.
+#' 
+#' \code{\link{MoE_control}} allows supplying additional arguments to \code{\link{MoE_clust}} and \code{\link{MoE_stepwise}} which govern, among other things, controls on the inclusion of an additional noise component and controls on the initialisation of the allocations for the EM/CEM algorithm.
 #'
 #' A dedicated plotting function exists for visualising the results using generalised pairs plots, for examining the gating network, and/or log-likelihood, and/or clustering uncertaines, and/or graphing model selection criteria values. The generalised pairs plots (\code{\link{MoE_gpairs}}) visualise all pairwise relationships between clustered response variables and associated continuous, categorical, and/or ordinal covariates in the gating &/or expert networks, coloured according to the MAP classification, and also give the marginal distributions of each variable (incl. the covariates) along the diagonal.
 #'
@@ -14,6 +16,8 @@
 #'
 #' Finally, a \code{predict} method is provided for predicting the fitted response and probability of cluster membership (and by extension the MAP classification) for new data, in the form of new covariates and new response data, or new covariates only.
 #'
+#' Other functions also exist, e.g. \code{\link{MoE_crit}}, \code{\link{MoE_dens}}, \code{\link{MoE_estep}}, \code{\link{MoE_compare}}, and \code{\link{aitken}}, which are all used within \code{\link{MoE_clust}} but are nonetheless made available for standalone use. 
+#' 
 #' The package also contains two data sets: \code{ais} and \code{CO2data}.
 #'
 #' @section Details:
@@ -37,21 +41,43 @@
 #' @examples
 #' \dontrun{
 #' data(ais)
-#' res <- MoE_clust(ais[,3:7], G=2, gating=~BMI, expert=~sex,
-#'                  modelNames=c("EVE", "VVE", "VEE"), network.data=ais)
-#' plot(res, what="gpairs")
+#' 
+#' # Fit two models
+#' res1  <- MoE_clust(ais[,3:7], G=2, gating=~BMI, expert=~sex,
+#'                    modelNames=c("EVE", "VVE", "VEE"), network.data=ais)
+#' res2  <- MoE_clust(ais[,3:7], G=2, equalPro=TRUE, expert=~sex,
+#'                    modelNames=c("EVE", "VVE", "VEE"), network.data=ais) 
+#'         
+#' # Compare the best model from each set of results
+#' (comp <- MoE_compare(res1, res2, optimal.only=TRUE))
+#' 
+#' # Procude a plot for the optimal model                                                   
+#' plot(comp$optimal, what="gpairs")
 #'
 #' data(CO2data)
 #' CO2   <- CO2data$CO2
 #' GNP   <- CO2data$GNP
 #' 
+#' # Fit a range of models 
 #' m1    <- MoE_clust(CO2, G=1:3)
 #' m2    <- MoE_clust(CO2, G=2:3, gating= ~ GNP)
 #' m3    <- MoE_clust(CO2, G=1:3, expert= ~ GNP)
 #' m4    <- MoE_clust(CO2, G=2:3, gating= ~ GNP, expert= ~ GNP)
 #' m5    <- MoE_clust(CO2, G=2:3, equalPro=TRUE)
 #' m6    <- MoE_clust(CO2, G=2:3, expert= ~ GNP, equalPro=TRUE)
-#' MoE_compare(m1, m2, m3, m4, m5, m6)}
+#'
+#' # Extract the model with highest BIC
+#' (comp <- MoE_compare(m1, m2, m3, m4, m5, m6, criterion="bic"))
+#'  
+#' # See if a better model can be found using greedy forward stepwise selection
+#' # Conduct a stepwise search on the same data
+#' (mod1 <- MoE_stepwise(CO2, CO2data[,"GNP", drop=FALSE]))
+#' 
+#' # Conduct another stepwise search considering models with a noise component
+#' (mod2 <- MoE_stepwise(CO2, CO2data[,"GNP", drop=FALSE], noise=TRUE))
+#' 
+#' # Compare all sets of results to choose the optimal model
+#' (best <- MoE_compare(mod1, mod2, comp, pick=1)$optimal)}
 #' @docType package
 #' @keywords package
 "_PACKAGE"