+++ R/data.R
@@ -21,7 +21,7 @@
 #' @references Cook, R. D. and Weisberg, S. (1994), \emph{An Introduction to Regression Graphics}. John Wiley & Sons, New York.
 #' @examples
 #' data(ais, package="MoEClust")
-#' pairs(ais[,c(3:4, 10:13)], col=as.numeric(ais[,1]), main = "AIS data")
+#' pairs(ais[,c(3:7)], col=as.numeric(ais$sex), main = "AIS data")
 #' @docType data
 #' @keywords datasets
 #' @usage data(ais)